build(opencascade-sys): rework building process of OpenCACADE
The previous building process had multiple issues:

- When for example running 'cargo clean && cargo build --workspace
  --release'. Cargo launched two simultaneous compilations of
  opencascade resulting in a race conditions cloning the source code.

- Builds and source code where compiled into the source directory
  instead of the target/ folder.

- Builds of OpenCASCADE did not reflect the correct optmimization level,
  since building happened inside a build-dependency instead of a
  dependency of occara

This new approach fixes all of the above listed issues.
