--- crates/opencascade-sys/.gitignore
@@ -1,2 +1,0 @@
-occt_source/
-occt_build/
