+++ crates/opencascade-sys/Cargo.toml
@@ -6,7 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-
-[build-dependencies]
 cmake = "0.1.50"
 walkdir = "2.4.0"
